Rename weapons.py -> projectiles.py
